fix misplacement of functions for generating rsa keys
